Remove sandbox /portal path (#2093)
## 🎫 Ticket

N/A

## 🛠 Changes

@jdettmannnava surfaced that the `/portal` path is used in the sandbox
to route back to the homepage. However, this conflicts with the routing
set up for dpc-portal.

This removes the /portal routing (which is just a duplicate of the root
/ homepage). Instead, we should explicitly route people to the root
path.

## ✅ Acceptance Validation

Unit tests, which have tests for redirection already.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.