[DPC-3970] Added invitation_id to ao_org_links (#2112)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3970

## 🛠 Changes

Added `invitation_id` column to `ao_org_links` table as a foreign key to
the `invitations` table.

## ℹ️ Context for reviewers

This ticket was to add `invitation_id` to `ao_org_links`, and
`invited_by_id` to `invitations`, but the `invitations` table had
already been previously updated so we only had to do `ao_org_links`.

## ✅ Acceptance Validation

Ran the migration locally and the new column was created as a foreign
key. All tests are passing.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.