[DPC-3981] Me/dpc 3981 fix smoke tests (#2138)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3981

## 🛠 Changes

Updates `make smoke/local` so that when run on Jenkins it purges
JMeter's utility library and rebuilds it based on dpc-smoketest's class
path.

## ℹ️ Context for reviewers

The original ask for this ticket was to fix failing DPC smoke tests in
Jenkins. They were failing because we had updated DropWizard, and in
that process bumped the jackson-annotations library from 1.13.3 to
2.16.1. This shouldn't have been a problem, but since dpc-smoketest is a
JMeter component, when run through Jenkins its dependencies are pulled
from JMeter's static utility library where they have to be manually
updated (See
[here](https://jmeter.apache.org/usermanual/get-started.html#classpath)
for more info). Despite updating our pom file with the new dependency,
the JMeter library still contained the old version. The fix was to
manually copy `jackson-annotations.2.16.1.jar` to the JMeter library and
delete `jackson-annotations.2.13.3.jar`.

This changes in this PR are to rebuild JMeter's static library on the
Jenkins server each time the smoke test is run so we won't have to worry
about this in the future. The process goes like this:

1. Check to see if the JMeter utility library directory exists.
2. Delete all of the *.jar files from the library.
3. Use Maven to copy all of dpc-smoketest's dependency jars into the
JMeter library.
4. Run as usual.

## ✅ Acceptance Validation

- Smoke tests are currently passing (See
[here](https://management.dpc.cms.gov/job/DPC%20-%20Smoke%20Test/)).
- A version of dpc-smoketest with logging turned up was run on Jenkins
to ensure the correct version of the jackson-annotations jar was being
loaded.
- When run on a system without a JMeter library, i.e. on your local
machine, everything still runs the same.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.