DPC-4028 Replace CPI API Gateway stubs with mock server (#2143)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4028

## 🛠 Changes

- FakeCpiGateway: sinatra app to emulate CPI API Gateway
- Use webmock to direct calls to IDM identity service and CPI API
Gateway to FakeCpiGateway
- Remove CpiApiGatewayClient doubles

## ℹ️ Context for reviewers

We want to use an internally-running Sinatra app to emulate the CPI API
Gateway in lower environments for a number of reasons. This step allows
us to test the integration before actually implementing it.

## ✅ Acceptance Validation

Doubles removed and tests pass.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.