[DPC-4029] Add jmacaroons library to local repository (#2147)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4029

## 🛠 Changes

This adds the `jmacaroons` library to a local repository, and includes
that repository in the relevant pom files.

## ℹ️ Context for reviewers

This library was being pulled from Jitpack, and is not hosted in Maven.
As we found the need to remove Jitpack as a remote repository, we
implemented a quick fix to put the `jmacaroons` jar in a local file, and
have the POM dependency use the `system` scope. This, however, violated
a rule in SonarQube, causing a build failure, and beyond that, is
generally considered a poor practice.

This solution presents a more stable fix to our issue.

## ✅ Acceptance Validation

- The build and all unit and integration tests pass via `make ci-app`.



## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.