Make local dev sidekiq more robust (#2162)
## 🎫 Ticket

No ticket

## 🛠 Changes

- Local portal-sidekiq loads GOLDEN_MACAROON if not present
- ActiveJob queue adapter set to :test in test environment

## ℹ️ Context for reviewers

- I noticed that ProviderOrganizations were not being created in the API
on create when developing locally. The sidekiq logs showed it did not
have permission to create the organization in the API. We already
optionally set the GOLDEN_MACAROON on boot for the rails apps, so I
ported the code to run before starting local portal sidekiq as well.
- I also noticed that some tests were calling redis when a provider
organization was created. Setting the queue adapter to :test fixed the
issue.

## ✅ Acceptance Validation

- Tests that made calls to sidekiq no longer do
- Making a provider organization locally creates the organization in the
API

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.