Me/dpc 4005 missing logs (#2164)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4005

## 🛠 Changes

- Set loggers for `DPCAuthFilter`, `GeneratedRequestIdFilter` and
`LogResponseFilter` to `INFO` so that contextual information is written
to our Splunk logs for easier debugging.
- Set default log levels to `WARN` instead of `ERROR`.

## ℹ️ Context for reviewers

We noticed that certain logs from the above classes were no longer
getting written out, and it was because they are logged at `INFO` and
our log levels were set to `ERROR`.

We talked about potentially setting the default log level to `INFO` in a
previous ticket (#2158), but the thought was that it might lead to
flooding our logs. As a bit of a compromise, I set them to WARN in this
ticket, but I'm open to changing them if anyone feels strongly about it.
If we find more missing logs we can always turn logging up to `INFO`,
and if the logs get a little too crowded we can turn them back to
`ERROR`.

## ✅ Acceptance Validation

Ran locally and ensured the logs are getting produced.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.