docs: Update portal README with manual workflow instructions (#2170)
## 🎫 Ticket

N/A

## 🛠 Changes

- Add scripts to generate GOLDEN_MACAROON as part of `make
portal-console` and `make portal-sh`
- Update README with instructions on how to go through Org setup and AO
invitation flow
- Reorganize README a bit to consolidate test instructions and such

## ℹ️ Context for reviewers

Wanted to go through the current workflow and had to trawl through the
code a bit to figure out AO invitations as it is currently built. Added
documentation while getting it to work.

## ✅ Acceptance Validation

Walked through the steps as written in order to get AO invitation
working for myself.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.