Add instructions for fixing ExpirationJobTest run in IntelliJ debugger (#2221)
## 🎫 Ticket

## 🛠 Changes

When attempting to run `ExpirationJobTest` from IntelliJ, using the
instructions in the README section `Debugging Integration Tests`, I was
getting a port collision error. This is because this particular test is
unique in that it runs in DropWizard as the `attribution` service. The
other integration tests run as the `dpc-api` service, and the
instructions were written with those in mind. I've added the case of
`ExpirationJobTest` to the README instructions for future use.

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->