Me/dpc 4166 remove hashed mbis (#2217)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4166

## 🛠 Changes

- Updated BBClient to so that patient searches by MBI are done with a
post instead of a get, and the MBI is sent in the post body.
- Removed MBI hashing since it's no longer necessary.
- Removed BBClientV2 and its config.

## ℹ️ Context

The BFD team asked us to begin sending MBIs via post. Instead of
updating BBClientV2 along with BBClient, this PR removes it since it's
not in use.

Ops changes: https://github.com/CMSgov/dpc-ops/pull/722

## 🧪 Validation

Ran locally and ensured we were able to pull data from BFD sandbox.