[DPC-4212] Unit test ExpireAttributions job (#2251)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4212

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
- Creates a set of unit tests to ensure that the ExpireAttributions job
correctly expires and deletes attribution records.
- Deletes the non-functional integration test class
- ~~Updates `rexml` in Rails libraries, to fix failing GitHub Actions~~
-- Fixed in a merged PR

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->
The old integration test class was non-functional after a recent upgrade
to the Dropwizard library, and seemingly wasn't testing the correct
functionality of the job. After some research, we decided to upgrade
this to a unit test.

The difficulty came in working out how to mock and persist data
temporarily. This PR represents an approach that works and passes.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->