[DPC-4212] Test error case in doJob (#2254)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4212

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
- Add a test for the error case in ExpireAttributions job.

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->
After merging the [previous
PR](https://github.com/CMSgov/dpc-app/pull/2251), CI failed due to a
SonarQube Quality Gate failure. The only case that isn't covered,
according to SonarQube, is the error case, so this PR should resolve the
issue.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->