[DPC-4259] Add lambda for syncing ip address set (#2269)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4259

## 🛠 Changes

Adds lambda to keep IP sets in sync on AWS.

## ℹ️ Context

By default, the production DPC API is not open to the public network. Organizations must provide IP addresses for inclusion in DPC's network allowlist so that they can access the DPC API.

We are currently storing these IP addresses in the database, but they are not automatically synced to the AWS WAF IP Set for functional purposes.

## 🧪 Validation

main_test.go includes an integration test that updates the WAF set,
verifies the update, and then resets it.