[DPC-4325] Update CD Invite mailer link (#2292)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4325

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
This updates a link in the cd invite mailer to be consistent with the
equivalent link in the ao invite mailer.

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->
I
[originally](https://github.com/CMSgov/dpc-app/pull/2291#discussion_r1805186111)
set this link to what was in the Confluence doc, but was made aware that
this is inconsistent from the AO mailer. Upon discussing with content,
we decided to make them consistent.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->
- [ ] Manual testing