[DPC-4213] Add error on create existing group (#2298)
## 🎫 Ticket


https://github.com/CMSgov/dpc-app/compare/me/dpc-4213-create-duplicate-rosters?expand=1

## 🛠 Changes

Attempting to create a roster that already exists for a provider in an
org will return an error. Previously, the request would be ignored, and
the already existing roster would be returned with a 200.

## ℹ️ Context

After much
[discussion](https://cmsgov.slack.com/archives/CJL0FRN2E/p1722519081409839),
I went with the majority and made the error a 403. If we change our
minds before this is merged it's only a two line change.

## 🧪 Validation

Ran locally and verified.