[DPC-4246] Disable default logging in opt-out-import lambda (#2313)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4246

## 🛠 Changes

Removed default logging from AWS SDK Golang config.

## ℹ️ Context

The AWS SDK (Golang, v2) automatically logs errors to stderr in a
non-JSON format that is not parseable by Splunk. This is not necessary
since our application itself will catch and log errors in the
standardized format, with the standardized log information.

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->