Use ssl to connect to databases in lambdas (#2334)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4415

## 🛠 Changes

Conditionally force sslmode depending on environment

## ℹ️ Context
In our upgrade to postgres 16, we have to use ssl. We had ssl turned off
for our lambdas so they could run against our local database (for
testing), which are not set up for ssl. New code runs in sslmode except
when 'testing' (which includes running locally)

## 🧪 Validation
Tested both locally (test and run-local)
Uploaded export to test and it worked.
Uploaded import to test but unsure how to test?