[DPC-4400] Load testing infrastructure (#2343)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4400

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
This PR creates a new module that we can use to develop load tests for
DPC, as well as a GitHub Action to run it manually. Currently, the
script is a proof of concept that calls a test endpoint provided by
Grafana (i.e. the maintainer of K6).

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->
This sets up the underlying infrastructure needed to start developing
our load tests.

Follow-up work is needed to unblock development. See the following
tickets:
https://jira.cms.gov/browse/DPC-4401
https://jira.cms.gov/browse/DPC-4427

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->
- [x] New GitHub Action runs successfully:
https://github.com/CMSgov/dpc-app/actions/runs/12202130538/job/34042108472