DPC-4349 Log AO Invitations asynchronously (#2347)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4349

## 🛠 Changes
- Added LogAoInviteJob to Log using Sidekiq
- Use LogAoInviteJob instead of Rails.logger in AoInvitationService

## ℹ️ Context

We use a rake task run on a specially-deployed Fargate task to generate
invitations to AOs. Despite following instructions, we have not been
able to configure these tasks to log to CloudWatch. Therefore, the
creation of our AO invitations are not being logged.
This workaround sends the log messages to our sidekiq instance to log,
which works.

## 🧪 Validation

Automated testing.
Deployed branch to dev, generated AO Invitation via rake tasks, and
found record in Splunk.