[DPC-4427] Fetch bearer token for load testing (#2356)
## 🎫 Ticket

https://jira.cms.gov/browse/4427

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->

- Adds environment variables necessary to generate a JWT for
authentication with DPC API.
- Fetches a bearer token for DPC API. 
- Adds documentation.

## ℹ️ Context

In order to enable load testing the DPC API, we need an organization in
the test environment that can authenticate. This PR validates that this
new organization is functional; it relies on the organization's client
token, public key ID, and RSA private key, which are stored in secrets,
and accessed by the GitHub Actions workflow in question.

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->
- [ ] Bearer token is fetched successfully in GitHub Actions workflow:
https://github.com/CMSgov/dpc-app/actions/runs/12380543433/job/34557163479?pr=2356
- [ ] Manual testing locally; bearer testing is fetched successfully on
developer's own machine.