DPC-4433 Tag Repository (#2358)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4433

## 🛠 Changes

- Added utility workflow tag_release

## ℹ️ Context

We are migrating from Jenkins to github actions for deployment. The
first application to implement this functionality is the static site.
Part of deployment involves tagging releases, which all releases
require. I put the utility for tagging releases in dpc-app because it
feels more central than dpc-static-site.

This workflow duplicates the tagging part of
[Jenkinsfile.release_static_site](https://github.com/CMSgov/dpc-ops/blob/main/jenkins_files/Jenkinsfile.release_static_site)

## 🧪 Validation

Tagging ran successfully:
https://github.com/CMSgov/dpc-static-site/actions/runs/12377126243/job/34545914298