[DPC-4477] Use admin URL generate-token endpoint for auth in load test (#2420)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4477

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
This updates our load testing authentication to use the admin
generate-token endpoint.

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->
Previously, we were using the regular /Auth endpoint, which requires
users to generate a JWT. That's a computationally expensive operation
that caused our load tests to hang for 20+ seconds at a time. This
solution avoids the need for computationally expensive operations.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

- [ ] Load test runs successfully, and successfully fetches a bearer
token using the API admin URL
https://github.com/CMSgov/dpc-app/actions/runs/13162096608/job/36733157663
<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->